VoxelDistance log warning

+++ shaders/settings.glsl
@@ -13,7 +13,7 @@ const int noiseTextureResolution = 512;
 
 const bool shadowHardwareFiltering1 = true;
 const int shadowMapResolution       = 2048; // [512 1024 1536 2048 3072 4096 6144 8192 10240 12288 16384]
-const int voxelDistance             = 32;
+const float voxelDistance             = 32.0;
 const float shadowDistance          = 192.0; // [8.0 16.0 32.0 48.0 64.0 80.0 96.0 112.0 128.0 144.0 160.0 176.0 192.0 208.0 224.0 240.0 256.0 320.0 384.0 512.0 768.0 1024.0 1280.0 1536.0]
 const float shadowDistanceRenderMul = 1.0;
 const float shadowIntervalSize      = 2.0;
